test: 添加日志

+++ ctool4j-parent/ctool4j-component/ctool4j-cache/src/main/java/com/c332030/ctool4j/cache/annotation/CCacheable.java
@@ -40,6 +40,6 @@ public @interface CCacheable {
      * 缓存过期时间，单位秒，为 0 则为永久
      * @return 缓存过期时间
      */
-    long expire() default 0;
+    int expire() default 0;
 
 }
